Add files via upload
Hello and welcome to the Guessing Game..!
This interactive game requires participants to predict a random number from a certain range. Players receive suggestions regarding whether their guess is too high or too low via a recursive function, encouraging them to the proper answer. The game constantly increases the range with each try, increasing the challenge. As you try to crack the code in as few trials as possible, put your intuition and reasoning to the test! 
Good luck, and have fun with the Guessing Game!
Thank U...!